+++ wazo_confd/plugins/incall/validator.py
@@ -1,5 +1,4 @@
 # Copyright 2016-2019 The Wazo Authors  (see the AUTHORS file)
-# Copyright (C) 2016 Proformatique Inc.
 # SPDX-License-Identifier: GPL-3.0-or-later
 
 from wazo_confd.helpers.destination import DestinationValidator